move(FServiceClient): protocol -> provider :recycle:
less complicated import chain this way :shrug:
